requirements.txt
anthropic==0.37.1
pdf2image==1.17.0
Pillow==11.0.0